Update mapping to run more fluidly.
To create fluid motion, the code was updated to animate map and marker positions. Instead of instantly jumping to new coordinates, the map now uses map.flyTo() for a smooth panning effect.

For the marker, a requestAnimationFrame() loop was introduced. This function continuously calculates small steps between the old and new GPS points, moving the marker frame-by-frame along the path for a seamless and responsive tracking experience on screen.

NOTE: Also re-cloned repository